Add placing capability
Read description

**Uses keyboard; not likely to work on mobile**

##Notes
This *will* move if you are typing in a menu, but not in an alert.
Non-shifted presses will behave *unpredictably* if there are multiple pixels.
Shift-hold presses will behave more predictably. Using multiple pixels is not supported.
The pixel has a state of "solid" and a density of 2. It might be able to move through liquids and gases according to this density.

##Controls##
WASD to move
Z to shock
X to explode
Hold shift to repeat action
Hold alt with movement keys to place the current element
* It cannot place itself
* You can also use Alt+H to place on the right
Q to reset keys (current action, shift status, alt status) if they get stuck