Humanitize: HOUSE BUILDING UPDATE!
To all you fellas saying humans should build houses, well now they can! When the leader gets enough wood, they'll eventually build a home. When they gather enough stone and more wood, they'll furnish the house with a furnace and a workbench. With that workbench, they'll use grass, stone and wood to make a pickaxe, which is when they'll be making mining holes if there is a nearby stone patch. 

Changelog:
-Humans build wooden houses
-Humans can make pickaxes
-Humans can mine for rock with pickaxes
-Humans can make furnaces and workbenches
-Humans can harvest wood from tree branches
-Humans can harvest fruit from plants