chaos.js
okay, so i made this and posted a screenshot of it on the discord, with no plans of releasing this. but then, people were like “you should publish this” for SOME reason. it just randomizes the position, rotation and scale of everything on the page with a 15 second delay, so you have time to uninstall it if you dont want it anymore (and im sure nobody would want this)