Update and rename mods/betaworldgen.js to worldgenlibrary.js
betaworldgen.js is no longer in the beta, and to differentiate itself more from the "worldgen_test.js" mod, it has taken the label of a library as it simplifies world generation allowing mod makers to use this to make new biomes, although it can be used for fun and not solely for mod making. it also includes ore vein generation and has a biome specific to the plants.js mod.