Update variety_mod.js
  This Sandboxels mod adds more things to the game. For example, it adds moss, which grows sideways until something blocks it. It also adds different versions of flies, including female flies, which create fly eggs, which then hatch into fly larvae.
The fly larva have a small chance of turning into either a male fly or a female fly. The male flies act like the flies that were already implemented into the game. The mod also adds disinfectant, which kills flies and fly larva, and converts zombies to humans. When zombies are close enough to a human, it converts the human into a zombie. It also adds strange cells, which do a random task every game tick, and has a small chance of creating something random. The mod also includes growth fluid, that allows fly eggs and fly larva to grow faster, and trash bags, which stores a list of items, and has a small chance of turning into a full garbage bag, which has a small chance of creating garbage juice.